css/image-fit/js/ua-stylesheet/img/003.html

<!doctype html>
<title>img default style, src</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<img src=../../support/opera.png>
<script>
test(function() {
  var e = document.body.firstChild;
  assert_equals(getComputedStyle(e, '').OObjectFit, 'auto');
  assert_equals(getComputedStyle(e, '').OObjectPosition, '50% 50%');
});
</script>
